website/requirements.txt
cryptography
django==3.0.8
django-allauth==0.42.0
django-oauth-toolkit==1.3.2
oauthlib==3.1.0
pyjwt==1.7.1
psycopg2==2.8.5